Feature/agent authentication

+++ common/src/main/java/edu/uoc/tfgmonitorsystem/common/model/service/IDbSequenceService.java
@@ -2,5 +2,5 @@ package edu.uoc.tfgmonitorsystem.common.model.service;
 
 public interface IDbSequenceService {
 
-    int generateDbSequence(String seqName);
+    long generateDbSequence(String seqName);
 }
